Rename js file and start with a capital
