+++ src/Frontend/Themes/Bootstrap/src/Js/Index.js
@@ -3,13 +3,12 @@ import Bootstrap from 'bootstrap-sass/assets/javascripts/bootstrap';
 
 /* Utilities imports */
 import SweetScroll from "sweet-scroll";
-import resizeFunction from './utilities/resize';
-/* eg. import {fancyBox} from './utilities/fancybox'; */
+import ResizeFunction from './utilities/Resize';
+/*import {Fancybox} from './utilities/Fancybox';*/
 
 /* Theme imports */
-/* eg. import tooltip from './theme/tooltip'; */
+/* eg. import tooltip from './Theme/Tooltip'; */
 
 /* Renders */
 new SweetScroll();
-new resizeFunction();
-
+new ResizeFunction();
