+++ src/Frontend/Themes/Bootstrap/src/Js/utilities/DistributeHeight.js
