+++ src/Frontend/Themes/Bootstrap/src/Js/utilities/Fancybox.js
@@ -1,0 +1,12 @@
+import 'fancybox/dist/js/jquery.fancybox.pack';
+
+export class Fancybox {
+
+    constructor() {
+        this.initFancybox();
+    }
+
+    initFancybox() {
+        $('.fancybox').fancybox();
+    }
+}
